# BotSharp.WebHost/wwwroot/swagger.json
{
  "swagger": "2.0",
  "host": "localhost:7500",
  "basePath": "/",
  "schemes": [ "http" ],
  "info": {
    "title": "Articulate API Documentation",
    "version": "0.0.1",
    "contact": { "name": "Smart Platform Group" }
  },
...
  "paths": {
...
    "/intent/{id}": {
      "get": {
        "summary": "Find a model instance by id from the data source",
        "operationId": "getIntentId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "put": {
        "summary": "Update attributes for a model instance and persist it into the data source",
        "operationId": "putIntentId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 37" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "delete": {
        "summary": "Delete a model instance by id from the data source",
        "operationId": "deleteIntentId",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/agent/{id}/intent": {
      "get": {
        "summary": "Find list of intents linked with a model instance specified by id",
        "operationId": "getAgentIdIntent",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "The index of the first element to return. 0 is the default start.",
            "name": "start",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Number of elements to return from start. All the elements are returned by default",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "String that will filter values to return only those intents with part of this filter in their names",
            "name": "filter",
            "in": "query"
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/domain/{id}/intent": {
      "get": {
        "summary": "Find list of intents linked with a domain",
        "operationId": "getDomainIdIntent",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "The index of the first element to return. 0 is the default start.",
            "name": "start",
            "in": "query"
          },
          {
            "type": "number",
            "description": "Number of elements to return from start. All the elements are returned by default",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "description": "String that will filter values to return only those intents with part of this filter in their names",
            "name": "filter",
            "in": "query"
          }
        ],
        "tags": [ "domain" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/entity/{id}/intent": {
      "get": {
        "summary": "Find a model instance by id from the data source",
        "operationId": "getEntityIdIntent",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the entity",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "entity" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/intent/{id}/postFormat": {
      "get": {
        "summary": "Find a post format by intent id from the data source",
        "operationId": "getIntentIdPostformat",
        "parameters": [
          {
            "type": "string",
            "description": "Id of the intent",
            "x-convert": { "trim": true },
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "post": {
        "summary": "Create a new instance of a post format for the intent and persist it into the data source",
        "operationId": "postIntentIdPostformat",
        "parameters": [
          {
            "type": "string",
            "description": "Id of the intent",
            "x-convert": { "trim": true },
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 28" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "put": {
        "summary": "Update attributes of the post format of the intent and persist it into the data source",
        "operationId": "putIntentIdPostformat",
        "parameters": [
          {
            "type": "string",
            "description": "Id of the intent",
            "x-convert": { "trim": true },
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 42" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "delete": {
        "summary": "Delete a webhook instance by id from the data source",
        "operationId": "deleteIntentIdPostformat",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/intent/{id}/scenario": {
      "get": {
        "summary": "Find a scenario by intent id from the data source",
        "operationId": "getIntentIdScenario",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "post": {
        "summary": "Create a new instance of a scenario for the intent and persist it into the data source",
        "operationId": "postIntentIdScenario",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 26" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "put": {
        "summary": "Update attributes of the scenario of the intent and persist it into the data source",
        "operationId": "putIntentIdScenario",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 40" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "delete": {
        "summary": "Delete a scenario instance by id from the data source",
        "operationId": "deleteIntentIdScenario",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/intent/{id}/webhook": {
      "get": {
        "summary": "Find a webhook by intent id from the data source",
        "operationId": "getIntentIdWebhook",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "post": {
        "summary": "Create a new instance of a webhook for the intent and persist it into the data source",
        "operationId": "postIntentIdWebhook",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 27" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "put": {
        "summary": "Update attributes of the webhook of the intent and persist it into the data source",
        "operationId": "putIntentIdWebhook",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 41" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      },
      "delete": {
        "summary": "Delete a webhook instance by id from the data source",
        "operationId": "deleteIntentIdWebhook",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/agent/{id}/domain/{domainId}/intent/{intentId}": {
      "get": {
        "summary": "Find an intent by id given a domain and an agent",
        "operationId": "getAgentIdDomainDomainidIntentIntentid",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "domainId",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "intentId",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/agent/{id}/domain/{domainId}/intent/{intentId}/scenario": {
      "get": {
        "summary": "Find the scenario related with an intent, for the given domain and agent",
        "operationId": "getAgentIdDomainDomainidIntentIntentidScenario",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "domainId",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "intentId",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/agent/{id}/domain/{domainId}/intent/{intentId}/webhook": {
      "get": {
        "summary": "Find the webhook related with an intent, for the given domain and agent",
        "operationId": "getAgentIdDomainDomainidIntentIntentidWebhook",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "domainId",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "intentId",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
    "/agent/{id}/domain/{domainId}/intent/{intentId}/postFormat": {
      "get": {
        "summary": "Find the post format related with an intent, for the given domain and agent",
        "operationId": "getAgentIdDomainDomainidIntentIntentidPostformat",
        "parameters": [
          {
            "type": "number",
            "description": "Id of the agent",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the domain",
            "name": "domainId",
            "in": "path",
            "required": true
          },
          {
            "type": "number",
            "description": "Id of the intent",
            "name": "intentId",
            "in": "path",
            "required": true
          }
        ],
        "tags": [ "agent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
    "/intent": {
      "post": {
        "summary": "Create a new instance of the model and persist it into the data source",
        "operationId": "postIntent",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": { "$ref": "#/definitions/Model 8" }
          }
        ],
        "tags": [ "intent" ],
        "responses": {
          "default": {
            "schema": { "type": "string" },
            "description": "Successful"
          }
        }
      }
    },
...
  },
...
}